+++ CHANGES.md
@@ -148,6 +148,7 @@
     -   Fix insert markup functions in consective case([GH-283][])
     -   Fix hide markup issue in markdown/gfm-view-mode([GH-468][])
     -   Fix bold regexp issue([GH-325][])
+    -   Fix italic and punctual character issue([GH-359][])
 
   [gh-171]: https://github.com/jrblevin/markdown-mode/issues/171
   [gh-216]: https://github.com/jrblevin/markdown-mode/issues/216
@@ -202,6 +203,7 @@
   [gh-340]: https://github.com/jrblevin/markdown-mode/issues/340
   [gh-349]: https://github.com/jrblevin/markdown-mode/issues/349
   [gh-350]: https://github.com/jrblevin/markdown-mode/pull/350
+  [gh-359]: https://github.com/jrblevin/markdown-mode/issues/359
   [gh-369]: https://github.com/jrblevin/markdown-mode/pull/369
   [gh-378]: https://github.com/jrblevin/markdown-mode/pull/378
   [gh-383]: https://github.com/jrblevin/markdown-mode/issues/383